Create a new component in Pages directory for RoomGithubConditions.
This is to ensure different components are properly separated with proper tests for each.
